refactor: revise folder structure
