+++ src/js/models/CartItem.js
@@ -1,4 +1,4 @@
-import App from "../App";
+import App from "../App.js";
 export default class CartItem {
   /**
    * @type {string} #productId
