+++ src/js/models/Product.js
