+++ src/js/renderers/abstract/Renderer.js
